refactor: update MetamodelService package to no.novari in relevant classes
