+++ src/main/java/no/fintlabs/provider/event/response/ResponseFintEventConsumer.java
@@ -8,9 +8,9 @@ import no.fintlabs.kafka.common.topic.pattern.ValidatedTopicComponentPattern;
 import no.fintlabs.kafka.event.EventConsumerConfiguration;
 import no.fintlabs.kafka.event.EventConsumerFactoryService;
 import no.fintlabs.kafka.event.topic.EventTopicNamePatternParameters;
-import no.fintlabs.metamodel.MetamodelService;
 import no.fintlabs.provider.config.KafkaConfig;
 import no.fintlabs.provider.event.request.RequestEventService;
+import no.novari.metamodel.MetamodelService;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.springframework.context.annotation.Bean;
 import org.springframework.kafka.listener.ConcurrentMessageListenerContainer;
